cloud/python/src/gateway/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
Flask==2.3.2
Flask-PyMongo==2.3.0
gridfs-fuse==0.1.2.dev1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
llfuse==1.5.0
MarkupSafe==2.1.3
pika==1.3.2
pymongo==4.4.1
requests==2.31.0
urllib3==2.0.4
werkzeug==2.3.7
zipp==3.16.2
